feat(camunda-plugin): remove `deploy-changed-only` flag for deployment
The flag is removed by `enable-duplicate-filtering`. Thus, we don't
create a new version of the process when all files are unchanged,
but we do deploy all files if any of them has been modified.
This fixes a problem when files bound to deployment were not
deployed due to only process being changed.