chore(app): simplify and fix secret filtering
We apply filtering during logging, to ensure no secrets are leaked.

Unfortunately the previous implementation was not side-effect free,
and would inadvertently change the connection options, when used.

This changes the behavior, but also ensures the filtering is SIMPLE
enough. When it comes to security (and many other things) simplicity
trumps.