feat: add resources-dir flag
This adds an option to set the resources dir to a specificy
directory independently of the location of the executable and
the user data directory.

Example of use (Camunda modeler bundled with plugins):
Assume you want to deliver a bundle of the camunda modeler with
a certain set of plugins and/or flags via the Nix package manager.

The Camunda modeler is started using the app.asar file, which means
the location of the executable is the place where electron is installed.
We do not want to install our plugins and/or flags there as it is
generic software.

We could set the user data dir to a place in the Nix-store where the
plugins and/or flags are. However, this does not work. The Nix-store
is read-only and the user data dir is also used to save the state of
the Camunda modeler.

Therefore this PR introduces an other option to set a dedicated resources
directory. It is in between the app data dir and the user data dir as
we want to be able to overwrite stuff set at the app data dir and at the
same time not restric the user in his ability to configure on the user level.

If you use Nix to bundle the Camunda modeler with some flags and
plugins you put those in the Nix-store a read-only directory
structure.